DB.Postgres

Description

Note: Postgres backend functions.

Inserting records into the database

insertJSONBOverwriteOnConflict :: forall conn. IConnection conn => SourceID -> FileID -> SubjectID -> Value -> conn -> IO ()

Stores a row in the JSON Value format (serialising into the JSONB type in Postres). If the SourceID and SubjectID alredy exist, the old record is overwritten.

insertEAV :: IConnection conn => UUID -> SourceID -> FileID -> SubjectID -> Text -> Text -> conn -> IO ()

Stores a row in the EAV triples format.

Given the Value

{
  "subject_id": "patient0",
  "age": 25.5,
  "address": {
    "street": "Middle",
    "house_no": 26,
    "city": "Towcester"
  }
}

we can generate the EAV tripples via flattenToEAV:

(
  60bcf653-fc8e-40dd-bb15-c332a99969da,
  [
    (60bcf653-fc8e-40dd-bb15-c332a99969da, "subject_id", "patient0"),
    (60bcf653-fc8e-40dd-bb15-c332a99969da, "age", "25.5"),
    (d53f0483-1107-4a8c-b544-02e8ed54778d, "parent_uuid", "60bcf653-fc8e-40dd-bb15-c332a99969da"),
    (d53f0483-1107-4a8c-b544-02e8ed54778d, "parent_attr", "address"),
    (d53f0483-1107-4a8c-b544-02e8ed54778d, "city", "Towcester"),
    (d53f0483-1107-4a8c-b544-02e8ed54778d, "street", "Middle"),
    (d53f0483-1107-4a8c-b544-02e8ed54778d, "house_no", "26")
  ]
)

These are then inserted into the eavs table:

uid 	 source_id 	 fileName 	 subject_id 	 attribute 	 value

60bcf653-fc8e-40dd-bb15-c332a99969da 	 ... 	 ... 	 patient0 	 subject_id 	 patient0

60bcf653-fc8e-40dd-bb15-c332a99969da 	 ... 	 ... 	 patient0 	 age 	 25.5

d53f0483-1107-4a8c-b544-02e8ed54778d 	 ... 	 ... 	 patient0 	 parent_uuid 	 60bcf653-fc8e-40dd-bb15-c332a99969da

d53f0483-1107-4a8c-b544-02e8ed54778d 	 ... 	 ... 	 patient0 	 parent_attr 	 address

d53f0483-1107-4a8c-b544-02e8ed54778d 	 ... 	 ... 	 patient0 	 city 	 Towcester

d53f0483-1107-4a8c-b544-02e8ed54778d 	 ... 	 ... 	 patient0 	 street 	 Middle

d53f0483-1107-4a8c-b544-02e8ed54778d 	 ... 	 ... 	 patient0 	 house_no 	 26

Inserting record metadata (only for JSONB) into the database

insertJSONBAttributesValuesMergeOnConflict :: forall conn. IConnection conn => SourceID -> Value -> Value -> conn -> IO ()

The eavs_jsonb_attributes_values table stores all the possible values, generated by createAllPathsWithValues for all the inserted objects. For example, given these two rows of data:

[
  {"gender": "female", "age": 27},
  {"gender": "male", "age": 25}
]

We generate and insert the following records into eavs_jsonb_attributes_values:

source_id 	 attribute 	 values

... 	 {"gender": "str"} 	 ["female", "male"]

... 	 {"age": "int"} 	 [27, 25]

If we add a third record r

{"gender": "unknown", "age": 35},

and insert the result createAllPathsWithValues r into eavs_jsonb_attributes_values, the values of both rows will get merged with the values already in the table:

source_id 	 attribute 	 values

... 	 {"gender": "str"} 	 ["female", "male", "unknown"]

... 	 {"age": "int"} 	 [27, 25, 35]

cleanupJSONBAttributesValues :: forall conn. IConnection conn => conn -> IO (Maybe Int)

This function removes duplicates in the values column of the eavs_jsonb_attributes_values table.

Should be run after we have inserted all the records via insertJSONBAttributesValuesMergeOnConflict.

Storing error log output

insertError :: forall conn. IConnection conn => SourceID -> FileID -> String -> conn -> IO ()

clearErrors :: forall conn. IConnection conn => SourceID -> FileID -> conn -> IO ()

Misc

getFileID :: forall conn. IConnection conn => SourceID -> String -> conn -> IO (Maybe FileID)

deleteSourceFromEAVSJSONB :: forall conn. IConnection conn => SourceID -> conn -> IO ()

deleteSourceFromEAVSJSONBAttributesValues :: forall conn. IConnection conn => SourceID -> conn -> IO ()

Orphan instances

Convertible SqlValue Value

Instance details

Methods

safeConvert :: SqlValue -> ConvertResult Value

Convertible SqlValue FileID

Instance details

Methods

safeConvert :: SqlValue -> ConvertResult FileID

Convertible Value SqlValue

Instance details

Methods

safeConvert :: Value -> ConvertResult SqlValue

Convertible UUID SqlValue

Instance details

Methods

safeConvert :: UUID -> ConvertResult SqlValue

Convertible FileID SqlValue

Instance details

Methods

safeConvert :: FileID -> ConvertResult SqlValue

Convertible SubjectID SqlValue

Instance details

Methods

safeConvert :: SubjectID -> ConvertResult SqlValue

Convertible SourceID SqlValue

Instance details

Methods

safeConvert :: SourceID -> ConvertResult SqlValue