JSON.Utils

Description

Documentation

getSubjectID :: MonadThrow m => Value -> m SubjectID

createAllPathsWithValues :: Value -> HashMap Value (HashSet Value)

This function takes a JSON object and generates all paths to the leaf nodes, collecting all the leaf values in a set. E.g. given:

{
  "name": "Sam",
  "age": 25.5,
  "address": {
    "street": "Middle",
    "house_no": 26,
    "city": "Towcester"
  }
}

createAllPathsWithValues produces the following map:

{"name": "str"} -> {"Sam"},
{"age": "float"} -> {25.5},
{"address": {"street": "str"}} -> {"Middle"},
{"address": {"house_no": "int"}} -> {26},
{"address": {"city": "str"}} -> {"Towcester"}

We return a set of value, in case we encounter the same path at in different objects within the same array:

[
  {"name": "Alice", "age": 27}
  {"name": "John", "age": 25},
]

Here, createAllPathsWithValues outputs:

[{"name": "str"}] -> {"Alice", "John"},
[{"age": "int"}] -> {27, 25},

flattenToEAV :: MonadIO m => Value -> m (UUID, [(UUID, Text, Text)])

This function takes a JSON object and flattens it into a list of EAV triples. The entities are given a uniquely generated UUID, to prevent any clashes with data already in the EAV database. As a result, running this function repeatedly will generate different output, since the UUIDs are generated at random.

For example, given:

{
  "name": "Sam",
  "age": 25.5,
  "address": {
    "street": "Middle",
    "house_no": 26,
    "city": "Towcester"
  }
}

flattenToEAV returns the UUID of the root object and a list of triples, encoding the original object:

(
  60bcf653-fc8e-40dd-bb15-c332a99969da,
  [
    (60bcf653-fc8e-40dd-bb15-c332a99969da, "name", "Sam"),
    (60bcf653-fc8e-40dd-bb15-c332a99969da, "age", "25.5"),
    (d53f0483-1107-4a8c-b544-02e8ed54778d, "parent_uuid", "60bcf653-fc8e-40dd-bb15-c332a99969da"),
    (d53f0483-1107-4a8c-b544-02e8ed54778d, "parent_attr", "address"),
    (d53f0483-1107-4a8c-b544-02e8ed54778d, "city", "Towcester"),
    (d53f0483-1107-4a8c-b544-02e8ed54778d, "street", "Middle"),
    (d53f0483-1107-4a8c-b544-02e8ed54778d, "house_no", "26")
  ]
)

Nesting information is preserved by including the parent_uuid and parent_attr triples:

(<child object UUID>, "parent_uuid", <parent UUID>),
(<child object UUID>, "parent_attr", <attribute name>)