Runtime

Description

The runtime module exports the main functionality of the cv-convert tool.

Documentation

loadLibrary :: (MonadThrow m, MonadIO m, MonadReader JSContextPtr m) => LibFunctions -> m ()

The loadLibrary function takse a LibFunctions parameter, which either contains an inline JS script or points to an external script via URL. The library functions must be stored in a specific way, namely the library file must be of the form

Lib.fun1 = (a,b) => {...},
Lib.fun2 = ...

convertRow

Arguments

:: (MonadMask m, MonadIO m)

=> Int
Row index/counter

-> row
Row data

-> header
Row header

-> (row -> header -> m Value)
Coverter function taking the row and header returning a JSON Value

-> (Value -> [ValidatorFailure])
Validator function, used to check that the output of the converter function conforms to the JSON schema, specified in jsonSchema

-> DataOutput
Either a Database connection or a file handle, depending on where we output data

-> ErrorHandling

-> Value
Row data as JSON

-> m (HashMap Value (HashSet Value))
returns a map from JSON Values to a set of Values, by calling createAllPathsWithValues on the converter function output

This function runs the converter function on the row and header input data, then runs the validator(to check that the JSON Value that was produced validates agains the JSON schema).

If the output is valid, the data is either stored in a database or a JSON text file. When inserting into database, the output is stored in two different tables:

The insertJSONBOverwriteOnConflict function stores the data as Postgres JSONB in the eavs_jsonb_attributes_values table

flattenToEAV is used to generate flattened EAV triples form the given JOSn object, which are then stored in the eavs table via insertEAV

Finally, we return result of running createAllPathsWithValues on the JSON output data.

processRow

Arguments

:: (FoldableWithIndex Int f, MonadThrow m)

=> f row
Any list like data structure containing rows we can loop over with an index

-> (Int -> row -> m (HashMap Value (HashSet Value)))
Function that consumes the row together with it's index

-> m (HashMap Value (HashSet Value))

Helper function which loops over the rows of the given input, collecting and merging all the resulting maps (generated inside convertRow via createAllPathsWithValues)

processTxtFile :: (MonadMask m, MonadIO m, MonadReader JSContextPtr m) => (JSValue -> JSValue -> m Value) -> (Value -> [ValidatorFailure]) -> FilePath -> DataOutput -> ErrorHandling -> Int -> m (HashMap Value (HashSet Value))

processXlsxFile :: (MonadMask m, MonadIO m, MonadReader JSContextPtr m) => (JSValue -> JSValue -> m Value) -> (Value -> [ValidatorFailure]) -> FilePath -> DataOutput -> ErrorHandling -> Maybe SheetName -> m (HashMap Value (HashSet Value))

processCsvFile :: (MonadMask m, MonadIO m, MonadReader JSContextPtr m) => (JSValue -> JSValue -> m Value) -> (Value -> [ValidatorFailure]) -> FilePath -> DataOutput -> ErrorHandling -> m (HashMap Value (HashSet Value))

processJsonFile :: (MonadMask m, MonadIO m, MonadReader JSContextPtr m) => (JSValue -> JSValue -> m Value) -> (Value -> [ValidatorFailure]) -> FilePath -> DataOutput -> ErrorHandling -> m (HashMap Value (HashSet Value))

processFile

Arguments

:: (MonadMask m, MonadIO m, MonadReader JSContextPtr m)

=> (JSValue -> JSValue -> m Value)
Function taking the row and header JSValues, returning a JSON Value

-> (Value -> [ValidatorFailure])
Validator function, used to check that the output of the converter function conforms to the JSON schema, specified in jsonSchema

-> FilePath
Path of the input file

-> FileType
Number of rows to skip/row number to start from only used when FileType is TXTFile. Default is 0. Optional sheet name only used when FileType is XLSXFile.

-> DataOutputOpt

-> ErrorOpt
Describes the error behaviour when parsing a row of the input.

-> m ()