Runtime.Types

Documentation

data FileType

Type encoding the input document type along with potential parsing/processing options.

Constructors

TXTFile

Fields

startFromLine :: Int

JSONFile

CSVFile

XLSXFile

Fields

sheetName :: Maybe SheetName

data Specified a

Constructors

Specified a

Inferred a

Unspecified

parseFileType :: Maybe Int -> Maybe SheetName -> Maybe Text -> Parser (Specified FileType)

Parses a file extension, such as .txt into TXTFile. Returns Nothing if the extension does not mach any of the FileTypes.

fromSpecifiedFileType :: String -> Specified FileType -> FileType

data ErrorOpt

Data type used to describe the error reporting/handling behaviour when processing a row. Other than 'Terminate, all other options result in the error from a row being recorded/printed, with the computation continuing onto the next row.

Constructors

Terminate

LogToConsole

LogToFile

Log

newtype SheetName

Constructors

SheetName

Fields

unSheetName :: Text

data LibFunctions

Constructors

Inline ByteString

External

Fields

url :: Text

hash :: ByteString

data Settings

Internal representation of a '.settings' file.

Constructors

Settings

Fields

processFunction :: Text

A string containg the JS function which will be applied to the input

libraryFunctions :: Maybe LibFunctions

jsonSchema :: Maybe Schema

A JSON schema used for output validation

openAs :: Specified FileType

This parameter can be used to specify the parsing behaviour for files such as .phenotype, which should be parsed as JSON. If left blank, parsing defaults to either file extension (if it corresponds to one of the FileTypes), otherwise TXTFile. , startFrom :: Maybe Int -- ^ This parameter specifies how many rows should be skipped. Only works when parsing TXTFile files

onError :: Maybe ErrorOpt

Specifies the error logging behaviour , worksheet :: Maybe SheetName -- ^ Used to specify which worksheet should be parsed. Only works for XLSXFile files. If left blank, defaults to first found worksheet.

data DataOutputOpt

Defines the output behaviour, either writing to a JSON/SQL file or a database.

Constructors

JSONFileOutputOpt

DBOutputOpt ConnInfo SourceID
If writing to a database, the ConnInfo and SourceID paramameters must be supplied

SQLFileOutputOpt SourceID
If writing to an SQL file, the SourceID paramameter must be supplied

data DataOutput

Sum type for the handles for different output options.

Constructors

JSONFileOutput Handle
Handle points to an open JSON file

forall con_t.DBConn con_t => DBOutput

Fields

con :: con_t

Database connection handle, depends on the backend (MySQL/Postgres).

sourceID :: SourceID

fileID :: FileID

SQLFileOutput SourceID Handle
Handle points to an open SQL file

newtype ErrorHandling

Constructors

ErrorHandling (ErrorOpt, Maybe DataOutput)